fix linters error
