+++ modules/hi.js
@@ -1,0 +1,3 @@
+export default class hello {
+
+}
